moved ConsoleApp.sln to src folder

+++ src/BL/WeatherForecast.cs
@@ -11,11 +11,10 @@ namespace BL
     public class WeatherForecast
     {
         private static string apiKey = "c1c4d772f711221a4a1df5be101bb4a5";
-        private static string url1 = "https://api.openweathermap.org/data/2.5/weather?q=";
-        //private static string city = String.Empty;
+        private static string url1 = "https://api.openweathermap.org/data/2.5/weather?q=";        
         private static string url2 = "&appid=";
         private static string units = "&units=metric";
-        public static async Task<float> GetTemperature(string city)
+        public static async Task<double> GetTemperature(string city)
         {
             string finalUrl = $"{url1}{city}{url2}{apiKey}{units}";
             using (var client = new HttpClient())
@@ -27,7 +26,7 @@ namespace BL
                     var json = result.Content.ReadAsStringAsync().Result;
                     JObject obj = JsonConvert.DeserializeObject<JObject>(json);
                     JObject mainObj = obj["main"] as JObject;
-                    float temperature = (float)mainObj["temp"];
+                    double temperature = (double)mainObj["temp"];
                     return temperature;
                 }
                 else
@@ -36,7 +35,7 @@ namespace BL
                 }
             }
         }
-        public static async Task<string> Instructions(float temperature)
+        public static async Task<string> Instructions(double temperature)
         {
             string result = temperature switch
             {
